refactor: extract icons
